api/requirements.txt
fastapi
uvicorn[standard]
pydantic>=1.10,<3
rapidfuzz
httpx
python-dateutil
